requirements.txt
PyQt6>=6.3.1
Pyserial>=3.5
pyqt6-tools>=6.4.2.3.3
pyinstaller>=6.3.0